Add flow-covered packages section
Issue #4 requested a listing of packages which use Flow and export them for the user to benefit from.  This adds our flow-covered packages.  As we find 3rd party packages which have them we will try to submit pull requests for those as well. 

Hopefully this gets people started with submitting theirs as right now it is difficult to find libraries which provide this awesome feature of Flow.